% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolBioNet_Checker.R
\name{BoolBioNet_Checker}
\alias{BoolBioNet_Checker}
\title{Check the correctness of the Boolean network configuration}
\usage{
BoolBioNet_Checker(OneBioNet)
}
\arguments{
\item{OneBioNet}{A four-element formatted list that records information of a Boolean network.}
}
\value{
A logical value. If \code{TRUE}, it enables further downstream analysis. Abnormal results or invalid inputs will generate error messages to assist with troubleshooting.
}
\description{
The function evaluates the lengths of node names, input and output edges, and Boolean function configurations to assess the structural validity of the Boolean network.
}
\details{
All information in the network is mutually consistent. For instance, the number of Boolean functions matches the number of nodes; if a node has k inputs, it implies that the node has a k-bit truth table (or Boolean function), etc. See \link{BoolGRN_CellCollective}.
}
\examples{
# BoolBioNet_Checker(BoolGRN_CellCollective$c_1557);
# Return TRUE (Denote all information are appropriate)
BoolBioNet_Checker(BoolGRN_CellCollective$c_1557)
}
