% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolBioNet_FBLoops.R
\name{BoolBioNet_FBLoops}
\alias{BoolBioNet_FBLoops}
\title{Calculate the network's feedback loop}
\usage{
BoolBioNet_FBLoops(
  OneNet,
  AnalyType = c("loop", "length", "edgeattr", "loopsign"),
  Isolated = TRUE
)
}
\arguments{
\item{OneNet}{A four-element formatted list that records information of a Boolean network.}

\item{AnalyType}{A string indicating the feature of the feedback loop being analyzed.
Only those "real" strongly connected components were used for analysis. 
(See \link{BoolBioNet_StrConComp})
\itemize{
  \item \code{loop}: Return the list of feedback loops. 
  \item \code{length}: Count the distribution of feedback loop lengths.
  \item \code{edgeattr}: Analyze the fundamental properties of each edge in the
loop, including the number of embedded loops, sign, edge activity, effectiveness.
  \item \code{loopsign}: Return the sign property of each cycle.
}
The sign attributes are categorized into three types: positive (+), negative 
(-), and hybrid (?). For the definitions of sign, edge activity, and effectiveness,
refer to the following papers: 
[\href{https://doi.org/10.1007/s11538-008-9304-7}{Aracena2008}],
[\href{https://doi.org/10.1073/pnas.0407783101}{Kauffman2004}],
[\href{https://doi.org/10.1073/pnas.2022598118}{Gates2021}].}

\item{Isolated}{A logical value. Should keep self-loops? (Default: \code{TRUE})}
}
\value{
Depending on the value of \code{AnalyType}, the following types of 
return values may result:
\itemize{
  \item \code{loop}: a list, with the names of the elements indicating the 
SCC composition and the feedback loop.
  \item \code{length}: a matrix storing the distribution of feedback loop lengths.
  \item \code{edgeattr}: a datamframe characterizing the properties of each 
edge in the loop.
  \item \code{loopsign}: a matrix recording the sign attributes of each edge 
within each cycle.
}
}
\description{
This function provides a brief analysis of feedback loops within 
the systems.
}
\examples{
# Analyze various properties of feedback loops in gene networks.
# Here is an example of the network [k_1753781] from Kadelka's paper.
# BoolBioNet_FBLoops(BoolGRN_KadelkaSet$k_1753781, "loop")
# [[1]]scc1_loop1
# [1] "A"  "T2" "T1" "B"  "A" 
# [[2]]scc1_loop2
# [1] "A"  "T2" "T1" "H"  "B"  "A" 
# ... ... 

# BoolBioNet_FBLoops(BoolGRN_KadelkaSet$k_1753781, "length")
#      Loop_Length Count
# [1,]           1     6
# [2,]           2     8
# ... ... 

# BoolBioNet_FBLoops(BoolGRN_KadelkaSet$k_1753781, "edgeattr")
#    from to support SignType activity effectiveness
# 1     A T1       4        - 0.015625    0.14955357
# 2     A T2       8        + 0.031250    0.09505208
# ... ... 

# BoolBioNet_FBLoops(BoolGRN_KadelkaSet$k_1753781, "loopsign")
#      Positive Negative Hybrid
# [1,]        3        1      0
# [2,]        4        1      0
# ... ... 

BoolBioNet_FBLoops(BoolGRN_KadelkaSet$k_1753781, "loop")
BoolBioNet_FBLoops(BoolGRN_KadelkaSet$k_1753781, "length")
BoolBioNet_FBLoops(BoolGRN_KadelkaSet$k_1753781, "edgeattr")
BoolBioNet_FBLoops(BoolGRN_KadelkaSet$k_1753781, "loopsign")

}
