% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolBioNet_SimplifiedNet.R
\name{BoolBioNet_SimplifiedNet}
\alias{BoolBioNet_SimplifiedNet}
\title{Simplify networks}
\usage{
BoolBioNet_SimplifiedNet(OneNet)
}
\arguments{
\item{OneNet}{a 4-element formatted list, record biological network information}
}
\value{
a 4-element formatted list, record the simplified network
}
\description{
All isolated/useless/ineffective nodes are removed from the systems 
(Especially after analyzing). Return a simplified network.
}
\examples{
# Analysis of one real Boolean genetic network. The following 
# "Analyzed_Net" and "Simplified_Net" respectively represents the simplified 
# or not after analyzing process. Compare them and understand the usage.
set.seed(1234)
Analyzed_Net=BoolBioNet_CoreDyn(BoolGRN_CellCollective$c_2084, AnalysisMod="scaling",
  ExternalSelfLoop=FALSE, ResidualNet = TRUE)[[4]];
Simplified_Net=BoolBioNet_SimplifiedNet(Analyzed_Net)
# Original network.
BoolBioNet_Visualization(BoolGRN_CellCollective$c_2084, "print")
# Anlayzed network.
BoolBioNet_Visualization(Analyzed_Net, "print")
# Simplified network.
BoolBioNet_Visualization(Simplified_Net, "print")

}
