% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolBioNet_SimpNet.R
\name{BoolBioNet_SimpNet}
\alias{BoolBioNet_SimpNet}
\title{Simplify networks}
\usage{
BoolBioNet_SimpNet(OneNet)
}
\arguments{
\item{OneNet}{A four-element formatted list that records information of a Boolean network.}
}
\value{
A four-element formatted list that records information of a Boolean network.
}
\description{
All isolated, redundant, or useless nodes are removed from the
system, particularly following the analysis of core dynamic components, 
resulting in a simplified network.
}
\examples{
# Analysis of a real Boolean genetic network. The terms "Analyzed_Net" and 
# "Simplified_Net" denote the network before and after simplification, respectively. 
# Compare the two to understand their differences and the purpose of the 
# simplification process.
set.seed(1234)
Analyzed_Net=BoolBioNet_CoreDyn(BoolGRN_CellCollective$c_2084, AnalysisMod="relevant",
  ExtSelfLoop=FALSE, ResidualNet=TRUE)[[4]]
Simplified_Net=BoolBioNet_SimpNet(Analyzed_Net)
# Original network.
BoolBioNet_Visualize(BoolGRN_CellCollective$c_2084, "print")
# Anlayzed network.
BoolBioNet_Visualize(Analyzed_Net, "print")
# Simplified network.
BoolBioNet_Visualize(Simplified_Net, "print")

}
