% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolBioNet_StrConComp.R
\name{BoolBioNet_StrConComp}
\alias{BoolBioNet_StrConComp}
\title{Caluclate network's strong connect component}
\usage{
BoolBioNet_StrConComp(OneNet)
}
\arguments{
\item{OneNet}{a 4-element formatted list, record Boolean network information.}
}
\value{
List, each element denotes one strong connect component.
}
\description{
This function briefly analyze the strong connect component for 
comparsion of core dynamic anslysis.
}
\examples{
# Analyze the strong connect componets of a network.
# Here show an example of net[c_2202] in Cell Collective.
# BoolBioNet_StrConComp(BoolGRN_CellCollective$c_2202)
# Return a List
# [[1]]
# [1] "Exocytosis"
# [[2]]
# [1] "Packaging_Proteins"
# [[3]]
# [1] "Protein_Phosphatase_1" "DARPP32"  "Calcineurin" 
# [4] "Calcium"               "Glutamate_Receptor"   
# ... ... 
BoolBioNet_StrConComp(BoolGRN_CellCollective$c_2202)
}
