% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolBioNet_Visualization.R
\name{BoolBioNet_Visualization}
\alias{BoolBioNet_Visualization}
\title{Print-out/Visualize networks}
\usage{
BoolBioNet_Visualization(
  RealBioNet,
  Outtype = c("print", "text", "dataframe", "igraph")
)
}
\arguments{
\item{RealBioNet}{a 4-element formatted list, record biological network information}

\item{Outtype}{Character, how the information will be presented? 
\code{print}: Display directly to the terminal; \code{text}: a string that 
save all information as \code{print} scenario; \code{dataframe}: a data.frame
that record "source-target" pairs; \code{igraph}: a igraph-object that visualize
the network (if package \code{igraph} is available).}
}
\value{
Specific formatted data of networks
}
\description{
Convert network into other standardized format or visualization 
[text.vector, data.frame, to.terminal or igraph.object (if available)].
}
\examples{
# Print the network of c_1969 in Cell Collective set.
# BoolBioNet_Visualization(CellCollective$c_1969,"print");
# Print Out to terminal
#  "PI3K-> PTEN-> [Akt] ->Cdc42_Rac1"
#  "cFOS-> cJUN-> [AP1] ->uPAR"
#  "JNK-> p38-> [ATF2] ->CyclinD ->PTGS2"
#               ... ...
BoolBioNet_Visualization(BoolGRN_CellCollective$c_1969,"print")

# BoolBioNet_Visualization(BoolGRN_CellCollective$c_1969,"text");
# Return a text vector
#                                Akt
# "PI3K-> PTEN-> [Akt] ->Cdc42_Rac1"
#                          AP1
# "cFOS-> cJUN-> [AP1] ->uPAR"
#          ... ...
BoolBioNet_Visualization(BoolGRN_CellCollective$c_1969,"text")

# BoolBioNet_Visualization(BoolGRN_CellCollective$c_1969,"dataframe");
# Return a data.frame
#         source        target
# 1         PI3K           Akt
# 2         PTEN           Akt
# 3         cFOS           AP1
#          ... ...
BoolBioNet_Visualization(BoolGRN_CellCollective$c_1969,"dataframe")

}
