% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolBioNet_Visualize.R
\name{BoolBioNet_Visualize}
\alias{BoolBioNet_Visualize}
\title{Print or visualize networks}
\usage{
BoolBioNet_Visualize(
  RealBioNet,
  Outtype = c("print", "text", "dataframe", "igraph")
)
}
\arguments{
\item{RealBioNet}{A four-element formatted list that records information of a
Boolean network.}

\item{Outtype}{A string indicating how the information will be structured or 
presented: 
\itemize{
\item \code{print}: Display directly to the terminal.
\item \code{text}: A string that stores all information in the same format as
 the \code{print} scenario.
\item \code{dataframe}: A data frame recording source-target pairs.
\item \code{igraph}: An igraph object used to visualize the network (if the 
\code{igraph} package is available).
}}
}
\value{
Specifically formatted network data.
}
\description{
Convert the network into alternative standardized formats or 
visualizations, such as \code{CharacterVector}, \code{data.frame}, terminal 
output, or \code{igraph} objects (if available).
}
\examples{
# Print the network [c_1969] from the CellCollective Set.
# BoolBioNet_Visualize(CellCollective$c_1969, "print");
# Print Out to terminal
#  "PI3K-> PTEN-> [Akt] ->Cdc42_Rac1"
#  "cFOS-> cJUN-> [AP1] ->uPAR"
#  "JNK-> p38-> [ATF2] ->CyclinD ->PTGS2"
#               ... ...
BoolBioNet_Visualize(BoolGRN_CellCollective$c_1969, "print")

# BoolBioNet_Visualize(BoolGRN_CellCollective$c_1969, "text");
# Return a text vector
#                                Akt
# "PI3K-> PTEN-> [Akt] ->Cdc42_Rac1"
#                          AP1
# "cFOS-> cJUN-> [AP1] ->uPAR"
#          ... ...
BoolBioNet_Visualize(BoolGRN_CellCollective$c_1969, "text")

# BoolBioNet_Visualize(BoolGRN_CellCollective$c_1969, "dataframe");
# Return a data.frame
#         source        target
# 1         PI3K           Akt
# 2         PTEN           Akt
# 3         cFOS           AP1
#          ... ...
BoolBioNet_Visualize(BoolGRN_CellCollective$c_1969,"dataframe")

}
