% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolFun_Complexity.R
\name{BoolFun_Complexity}
\alias{BoolFun_Complexity}
\title{Calculate the complexity of a Boolean function}
\usage{
BoolFun_Complexity(Bit_Vec)
}
\arguments{
\item{Bit_Vec}{A Boolean vector (or 0/1-valued \code{IntegerVector}) represents
the truth table of a Boolean function.}
}
\value{
A numeric value that denotes the complexity of a Boolean function.
}
\description{
The analysis depends on average of prime implicants in a Boolean 
function, based on Boolean Quine-McCluskey method.
}
\examples{
# Calculate the complexity of a Boolean function \code{c(0,0,1,0,0,1,1,0)}.
# Based on Quine-McCluskey method:
# (*10)+(101) ==> 1 
# (00*)+(*00)+(*11) ==> 0 
# Thus, 2+3 =5 items as the complexity.
# BoolFun_Complexity(c(0,0,1,0,0,1,1,0))
# Return 5
BoolFun_Complexity(c(0,0,1,0,0,1,1,0))

}
