% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolFun_NestedCana.R
\name{BoolFun_NestedCana}
\alias{BoolFun_NestedCana}
\title{Decode the structure of a nested canalized Boolean function}
\usage{
BoolFun_NestedCana(Bit_Vec, PrintOut = FALSE)
}
\arguments{
\item{Bit_Vec}{A Boolean vector (or 0/1-valued \code{IntegerVector}) represents
the truth table of a Boolean function.}

\item{PrintOut}{A logical value. Should output be displayed in the terminal? 
(Default: \code{FALSE}).}
}
\value{
A matrix that descirbes the nested structure of a canalized function.
}
\description{
A nested canalized Boolean function typically exhibits an "if-else
if..." regulatory structure. This function decodes and visually represents the
pattern to illustrate its logical hierarchy intuitively.
}
\examples{
# Show the canalized structure of a nested canalized function.
# set.seed(202501);
# aboolfun<-BoolFun_Generator("C", VarNum=7L, vars=c(4L,9L));
# df_boolfun<-BoolFun_NestedCana(aboolfun, PrintOut=TRUE);
# Return a matrix and display the nested relationships.
# if {x_5=1} ==> f(x)=0
# else if {x_5=0, x_6=1} ==> f(x)=0
# else if {x_5=0, x_6=0, x_7=1} ==> f(x)=0
# else if {x_5=0, x_6=0, x_7=0, x_3=1} ==> f(x)=1
# else if {x_5=0, x_6=0, x_7=0, x_3=0, x_4=1} ==> f(x)=0
# df_boolfun is:
#    x_1 x_2 x_3 x_4 x_5 x_6 x_7 Out
#  1  NA  NA  NA  NA   1  NA  NA   0
#  2  NA  NA  NA  NA   0   1  NA   0
#  3  NA  NA  NA  NA   0   0   1   0
#  4  NA  NA   1  NA   0   0   0   1
#  5  NA  NA   0   1   0   0   0   0

set.seed(202501)
aboolfun<-BoolFun_Generator("C", VarNum=7L, vars=c(4L,9L))
df_boolfun<-BoolFun_NestedCana(aboolfun, PrintOut=TRUE)

}
