% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolFun_Polynomial.R
\name{BoolFun_Polynomial}
\alias{BoolFun_Polynomial}
\title{Convert a Boolean function into threshold-based or polynomial forms}
\usage{
BoolFun_Polynomial(Bit_Vec, SpinLikeForm = FALSE, PolyForm = FALSE)
}
\arguments{
\item{Bit_Vec}{A Boolean vector (or 0/1-valued \code{IntegerVector}) represents the 
truth table of a Boolean function.}

\item{SpinLikeForm}{A logical value. Should the Boolean function be converted
into a spin-like form? (Default: \code{FALSE})}

\item{PolyForm}{A logical value. Should be converted according to a strict 
polynomial form.}
}
\value{
A three-element formatted list:
\itemize{
\item [[1]] Is Boolean satisfiability? (\code{TRUE} or \code{FALSE})
\item [[2]] The weight value of each variable and variable combination.
\item [[3]] The highest order of terms contained in the function.
}
}
\description{
Boolean functions are typically expressed as logical expressions
involving variables and logical operators such as AND, OR, and NOT. When 
represented in arithmetic form, a Boolean function corresponds to a unique 
polynomial, including higher-order terms. It should be noted that the term 
\emph{threshold-based} used here does not conform to the standard definition,
which usually refers to models restricted to first-order terms.
}
\details{
Boolean functions are typically represented in two standard forms:
the binary form (0/1) and the spin-like form (+1/-1), which are linearly 
interconvertible. \code{SpinLikeForm} allows users to obtain representations 
in the spin-like form, while \code{PolyForm} provides an option to output 
the strict polynomial representation of a Boolean function. For example, for 
the expression "\eqn{f=a|b}", the polynomial form is "\eqn{f=a+b+ab}", and 
the threshold form is "\eqn{f=(a+b)>0?}". See the provided examples for details.
In the output, variables (\eqn{x_1}, \eqn{x_2}, ... , \eqn{x_k}, for a 
k-input function) represent the input bits ordered from the low bit to the high bit.
}
\examples{
# Convert the a Boolean function into a polynomial expression:
# Case1: Boolean threshold
# BoolFun_Polynomial(c(0,1,1,0,1,0,0,0),F,F);
# Case2: Spin-like threshold
# BoolFun_Polynomial(c(0,1,1,0,1,0,0,0),T,F);
# Case3: Boolean strict polynomial form
# BoolFun_Polynomial(c(0,1,1,0,1,0,0,0),F,T);
# Case4: Spin-like strict polynomial form
# BoolFun_Polynomial(c(0,1,1,0,1,0,0,0),T,T);
# Return [[1]] sat:TRUE, [[2]] (see below), [[3]] HighOrder: 2 or 3
# Detail Weight of four types:
#         x_1    x_2    x_3  x_1x_2  x_1x_3  x_2x_3 x_1x_2x_3  thershold 
# Case1     1      1      1      -2      -2      -2      NULL          0 
# Case2    -2     -2     -2      -2      -2      -2      NULL          0 
# Case3     1      1      1      -2      -2      -2         3          0 
# Case4 -0.25  -0.25  -0.25   -0.25   -0.25   -0.25      0.75      -0.25

BoolFun_Polynomial(c(0,1,1,0,1,0,0,0), FALSE, FALSE) # Case1
BoolFun_Polynomial(c(0,1,1,0,1,0,0,0), TRUE,  FALSE) # Case2
BoolFun_Polynomial(c(0,1,1,0,1,0,0,0), FALSE, TRUE ) # Case3
BoolFun_Polynomial(c(0,1,1,0,1,0,0,0), TRUE,  TRUE ) # Case4

}
