% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolGRN_KadelkaSet.R
\docType{data}
\name{BoolGRN_KadelkaSet}
\alias{BoolGRN_KadelkaSet}
\title{91 genetic networks in Kadelka's paper}
\format{
An object of class \code{list} of length 91.
}
\usage{
data(BoolGRN_KadelkaSet)
}
\description{
A list containing networks from the paper, with each element 
representing an independent network. The number is consistent with the PMID 
of original reference. For example, \code{k_11082279} represents the paper 
"PMID:11082279".
}
\details{
Each element is a four-element formatted list that records information of 
a Boolean network (See \link{BoolGRN_CellCollective}). Detail information 
about dataset, please see original paper 
\href{https://doi.org/10.1126/sciadv.adj0822}{Kadelka2024}
}
\keyword{datasets}
