% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_BF_isPointed}
\alias{c_BF_isPointed}
\title{Determine whether a function belongs to a special type}
\usage{
c_BF_isPointed(boolfuns, k, Type, Showit)
}
\arguments{
\item{boolfuns}{[\code{LogicalVector}] the Boolean function logical vector.}

\item{k}{[\code{int}] the number of input variables of the function.}

\item{Type}{[\code{char}] function type.}

\item{Showit}{[\code{bool}] whether show the function?}
}
\description{
C++ prototype function of \link{BoolFun_Type}.
}
