% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_BF_Sensitivity}
\alias{c_BF_Sensitivity}
\title{Calculate the sensitivity of a Boolean function}
\usage{
c_BF_Sensitivity(boolfuns, k)
}
\arguments{
\item{boolfuns}{[\code{LogicalVector &}] the Boolean function logical vector.}

\item{k}{[\code{int}] the number of input variables of the function.}
}
\description{
C++ prototype function of \link{BoolFun_Sensitivity}.
}
