% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_BoolFun2Polynomial}
\alias{c_BoolFun2Polynomial}
\title{Convert a Boolean function into threshold-based or polynomial forms}
\usage{
c_BoolFun2Polynomial(VariableMat, MapTab, LogiSpin)
}
\arguments{
\item{VariableMat}{[\code{IntegerMatrix &}] a matrix of variable combination}

\item{MapTab}{[\code{IntegerMatrix &}] the truth table of Boolean function.}

\item{LogiSpin}{[\code{int}] Should show the detail polynomial form?}
}
\description{
C++ prototype function of \link{BoolFun_Polynomial}.
}
