% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_M_NestedCanalized}
\alias{c_M_NestedCanalized}
\title{Analyze the structure of a multi-valued nested canalized function}
\usage{
c_M_NestedCanalized(amulfun, k, L, Lens)
}
\arguments{
\item{amulfun}{[\code{IntegerVector &}] the function.}

\item{k}{[\code{int}] the number of input variables.}

\item{L}{[\code{int}] the level of function.}

\item{Lens}{[\code{int}] the length of mapping table (\eqn{L^k}).}
}
\description{
C++ prototype function of \link{MulVFun_is_NestedCana}
}
