% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_M_Signed}
\alias{c_M_Signed}
\title{Decode the information of a multi-valued signed function}
\usage{
c_M_Signed(amulfun, k, L, Lens)
}
\arguments{
\item{amulfun}{[\code{IntegerVector &}] the function.}

\item{k}{[\code{int}] the number of input variables of original function.}

\item{L}{[\code{int}] the level of original function.}

\item{Lens}{[\code{int}] the length of mapping table (\eqn{L^k}).}
}
\description{
C++ prototype function of \link{MulVFun_is_Signed}
}
