% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_MulF_Complexity}
\alias{c_MulF_Complexity}
\title{Calculate the complexity of a multi-valued function}
\usage{
c_MulF_Complexity(avec, k, L)
}
\arguments{
\item{avec}{[\code{IntegerVector}] the function.}

\item{k}{[\code{int}] the number of input variables.}

\item{L}{[\code{int}] the level of function.}
}
\value{
[\code{double}] the complexity of given function.
}
\description{
C++ prototype function of \link{MulVFun_Complexity}
}
