% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_MulV2Bool_Bool2MulV}
\alias{c_MulV2Bool_Bool2MulV}
\title{Boolean to multi-valued and multi-valued to Boolean transformation}
\usage{
c_MulV2Bool_Bool2MulV(OriMapTab, k, L, Thresholds, b2m)
}
\arguments{
\item{OriMapTab, }{[\code{IntegerVector}] original mapping table.}

\item{k, }{[\code{int}] the number of input variables.}

\item{L, }{[\code{int}] the level of multi-valued system.}

\item{Thresholds, }{[\code{IntegerVector}] thresholds for a multi-valued system.}

\item{b2m, }{[\code{int}] Is from Boolean to Multi-valued system? (>0, yes)}
}
\description{
C++ prototype function of \link{MulV2Bool_Bool2MulV}
}
