% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_Percolation_Simualtion}
\alias{c_Percolation_Simualtion}
\title{C++ prototype function for DNS_Percolation()}
\usage{
c_Percolation_Simualtion(
  sys_size,
  ll_system,
  sim_step,
  lat_type,
  obr_window,
  obf_type,
  bias_rf,
  obf_ratio,
  init_bias,
  net_type,
  net_f_para,
  obf_i_para1,
  obf_i_para2,
  OutPut,
  RuleType
)
}
\arguments{
\item{sys_size, }{system size}

\item{ll_system, }{level number of discrete system}

\item{sim_step, }{simulating steps}

\item{lat_type, }{lattice type}

\item{obr_window, }{Observe window}

\item{obf_type, }{ordered Boolean function type}

\item{bias_rf, }{biases in random function}

\item{obf_ratio, }{proportion of ordered Boolean function}

\item{init_bias, }{each}

\item{net_type, }{system topological type}

\item{net_f_para, }{topological configured parameters}

\item{obf_i_para1, }{configuration parameter for OBF (1)}

\item{obf_i_para2, }{configuration parameter for OBF (2)}

\item{OutPut, }{whehther output all states?}

\item{RuleType, }{update strategy}
}
\description{
C++ prototype function for DNS_Percolation()
}
\details{
Please see document of \link{DNS_Percolation}
}
