% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_ScalingLaw_Simualtion}
\alias{c_ScalingLaw_Simualtion}
\title{Analyze potential engaged nodes within Boolean/multi-valued system}
\usage{
c_ScalingLaw_Simualtion(
  sys_size,
  ll_system,
  obf_type,
  bias_rf,
  obf_ratio,
  net_type,
  net_f_para,
  obf_i_para1,
  obf_i_para2,
  PointedNode,
  PointValues,
  NodeDetailInfor,
  ReturnResidualNetwork
)
}
\arguments{
\item{sys_size}{[\code{int}] system size}

\item{ll_system}{[\code{int}] level number of discrete system}

\item{obf_type}{[\code{char}] ordered Boolean function type}

\item{bias_rf}{[\code{NumericVector}] biases in random function}

\item{obf_ratio}{[\code{double}] proportion of ordered Boolean function}

\item{net_type}{[\code{char}] system topological type}

\item{net_f_para}{[\code{char}] topological configured parameters}

\item{obf_i_para1}{[\code{int}] configuration parameter for OBF (1)}

\item{obf_i_para2}{[\code{int}] configuration parameter for OBF (2)}

\item{PointedNode}{[\code{IntegerVector}] which nodes should be fixed}

\item{PointValues}{[\code{IntegerVector}] pointed values of free/controlled nodes}

\item{NodeDetailInfor}{[\code{bool}] should return detail informations of node}

\item{ReturnResidualNetwork}{[\code{bool}] should return residual networks?}
}
\description{
C++ prototype function of \link{DNS_Engaged}
}
