% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNS_DamageSpread_Load.R
\name{DNS_DamageSpread_Load}
\alias{DNS_DamageSpread_Load}
\title{Analyze damage spread within system (Loaded systems)}
\usage{
DNS_DamageSpread_Load(
  RealBioNet,
  ExtSelfLoop = TRUE,
  Controller = NULL,
  ConVals = NULL,
  ExternalNode = NULL,
  Times = 1000L,
  Init_Dist = 0.1,
  Init_1_Ratio = c(0.5, 0.5),
  NumSys = 2L,
  UpdateRule = 1L
)
}
\arguments{
\item{RealBioNet}{A four-element formatted list that records information of a Boolean network.}

\item{ExtSelfLoop}{A logical value. Should inputs be converted into
self-loops? (Default: \code{TRUE})}

\item{Controller}{An integer or character vector indicating the nodes to be
controlled. (Default: \code{NULL})}

\item{ConVals}{A Boolean vector specifying the values of controlled of
controlled nodes. (Default: \code{NULL})}

\item{ExternalNode}{A Boolean vector specifying the states of non-input nodes.
(Default: \code{NULL}, indicating random assignment)}

\item{Times}{an integer, steps of simulating system.}

\item{Init_Dist}{a numeric value, initial normalized Hamming distance of two vector.}

\item{Init_1_Ratio}{a NumvericVector, proportion of each value (its length is 
\code{NumSys}).}

\item{NumSys}{an integer, number of discrete value.}

\item{UpdateRule}{an integer, update rule: 1=synchronous, 2=asynchronous, 
3=quick-asynchronous.}
}
\value{
A numeric value that representing the normalized Hamming distance.
}
\description{
Same as \link{DNS_DamageSpread}, but the network is loaded.
}
\details{
See \link{DNS_DamageSpread}.
}
\examples{
# set.seed(20250101L);
# DNS_DamageSpread(Size=1000L, SimStep=1000L, OBF_Type='C',
#   OBF_iPara1=1, OBF_iPara2=-1, OBF_Ratio=0.5);
# Return 0.335
# set.seed(20250101L)
# DNS_DamageSpread(Size=1000L, SimStep=1000L)

}
