% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulV2Bool_Bool2MulV.R
\name{MulV2Bool_Bool2MulV}
\alias{MulV2Bool_Bool2MulV}
\title{Boolean to multi-valued and multi-valued to Boolean transformation}
\usage{
MulV2Bool_Bool2MulV(
  aVec,
  k = 0,
  L = 3,
  Thres = NA,
  Bool2MulV = TRUE,
  MappingTable = FALSE
)
}
\arguments{
\item{aVec}{An IntegerVector that represents a mapping table of multi-valued function.}

\item{k}{An integer that denotes the number of input variables.}

\item{L}{An integer that represents the level of multi-valued system (Default: 3).}

\item{Thres}{An IntegerVector containing thresholds for a multi-valued system. 
Its length is \code{k+1}, with each element ranging from 1 to L-1. If not provided, 
the function randomly generate them (Default: \code{NA}).}

\item{Bool2MulV}{A logical value. Is the original function Boolean? If \code{TRUE} 
the transformation involves converting it from Boolean to multi-valued. 
Otherwise, from multi-valued to Boolean. (Default: \code{TRUE}).}

\item{MappingTable}{A logical value. Is the mapping table also returned? 
(Default: \code{FALSE})}
}
\value{
An IntegerVector (or IntegerMatrix if \code{MappingTable} is \code{TRUE}), 
denoting "mapping table" of target system.
}
\description{
Boolean to multi-valued and multi-valued to Boolean transformation
}
\examples{
# Transform a Boolean function to a multi-valued one.
# The Boolean is f(x,y)=x & y, a possible result see follows:
# set.seed(2025);
# MulV2Bool_Bool2MulV(c(0,0,0,1), k=2, L=3, 
#   c(1,1,1), <--- the two inputs and one output share same threholds:
#                  in this case,  0 mapto 0 and 1 mapto {1,2}
#   MappingTable = TRUE);
# Return
#       v1 v0 xx   #  Boolean system
#  [1,]  0  0  0   #    0  0  0 
#  [2,]  0  1  0   #    0  1  0 
#  [3,]  0  2  0   #    0  1  0 
#  [4,]  1  0  0   #    1  0  0 
#  [5,]  1  1  2   #    1  1  1 
#  [6,]  1  2  2   #    1  1  1 
#  [7,]  2  0  0   #    1  0  0 
#  [8,]  2  1  1   #    1  1  1 
#  [9,]  2  2  2   #    1  1  1 
set.seed(2025)
MulV2Bool_Bool2MulV(c(0,0,0,1), 2, 3, c(1,1,1), MappingTable=TRUE)

}
