% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulVFun_is_Domainted.R
\name{MulVFun_is_Domainted}
\alias{MulVFun_is_Domainted}
\title{Decode the information of a multi-valued domaineted function}
\usage{
MulVFun_is_Domainted(aVec, k, L, PrintOut = FALSE)
}
\arguments{
\item{aVec}{An IntegerVector that represents a mapping table of multi-valued 
function.}

\item{k}{An integer that denotes the number of input variables.}

\item{L}{An integer that represents the level of multi-valued system.}

\item{PrintOut}{A logical value. Should the output be displayed in the terminal?}
}
\value{
A list containing three elements: 
\itemize{
\item \code{[[1]] IsDomaint}: Is domainted? 1 for Yes, 0 for No.
\item \code{[[2]] Weight}: The weights of variables, ordered from low to high bits.
\item \code{[[3]] Delta}: The baselines of variables, ordered from low 
to high bits.}
}
\description{
The function analyzes and returns a list containing information
on whether the given function is domaineted and corresponding weights and 
baselines (deltas).
It is also compatible with Boolean system (L=2).
}
\examples{
# Show the information of a multi-valued linear threshold function.
# set.seed(1002);
# A_MulV_Fun=MulVFun_Generator("D", k=2L, L=3L, MappingTable=TRUE);
# View(A_MulV_Fun); # Show the mapping table.
# A_MulV_Fun_Domaint=MulVFun_is_Domainted(A_MulV_Fun[,3], 2, 3, TRUE);
# Terminal output: 
# f(~): {w_1=-5, w_0=3 | delta_0=-1, delta_1=0, delta_2=1}
# Above info means: avg max{i| \sum(w_i*x_i)+delta_i }
# A_MulV_Fun_Domaint # Show the information
# $IsDomaint
# [1] 1
# $Weight
# [1] 3 -5
# $Delta
# [1] -1  0  1

set.seed(1002)
A_MulV_Fun=MulVFun_Generator("D", k=2L, L=3L, MappingTable=TRUE)
A_MulV_Fun_Domaint=MulVFun_is_Domainted(A_MulV_Fun[,3], 2, 3, TRUE)

}
