% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_CheckValidBoolFun.R
\name{r_CheckValidBoolFun}
\alias{r_CheckValidBoolFun}
\title{Boolean function checker}
\usage{
r_CheckValidBoolFun(Bit_Vec)
}
\arguments{
\item{Bit_Vec}{A Boolean vector (or 0/1-valued IntegerVector) represents the 
truth table of a Boolean function.}
}
\value{
a List: [[1]] Boolean vector, [[2]] in-degree
}
\description{
Check whether the input vector is a valid Boolean function
}
