% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_GenerateTruthTable.R
\name{r_GenerateTruthTable}
\alias{r_GenerateTruthTable}
\title{Generate a truth table with pointed in-degree function.}
\usage{
r_GenerateTruthTable(inDegree = 2L, SpinLike = FALSE)
}
\arguments{
\item{inDegree}{an integrate value, represents a truth table of Boolean function.}

\item{SpinLike}{a logical, should output the spin-like form?}
}
\value{
a data.frame of input case enumeration
}
\description{
Generate a truth table with pointed in-degree function.
}
