% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transient_Process_A.R
\name{Transient_Process_A}
\alias{Transient_Process_A}
\title{Simulating roughly attractor analysis}
\usage{
Transient_Process_A(
  RealBioNet,
  SampleNum = 100L,
  ExtSelfLoop = TRUE,
  Controller = NULL,
  ConVals = NULL,
  ExternalNode = NULL,
  Times = 1000L,
  NumSys = 2L,
  UpdateRule = 1L
)
}
\arguments{
\item{RealBioNet}{A four-element formatted list that records information of a Boolean network.}

\item{SampleNum}{An integer pointing the number of randomly sampling initial states.}

\item{ExtSelfLoop}{A logical value. Should inputs be converted into
self-loops? (Default: \code{TRUE})}

\item{Controller}{An integer or character vector indicating the nodes to be
controlled. (Default: \code{NULL})}

\item{ConVals}{A Boolean vector specifying the values of controlled of
controlled nodes. (Default: \code{NULL})}

\item{ExternalNode}{A Boolean vector specifying the states of non-input nodes.
(Default: \code{NULL}, indicating random assignment)}

\item{Times}{an integer, steps of simulating system.}

\item{NumSys}{an integer, number of discrete value.}

\item{UpdateRule}{an integer, update rule: 1=synchronous, 2=asynchronous, 
3=quick-asynchronous.}
}
\value{
A matrix where each row represents the final state of a stochastic 
simulation, and each column represents a gene.
}
\description{
Like \link{DNS_DamageSpread}, but record final state.
}
\details{
See \link{DNS_DamageSpread}.
}
