% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transient_Process_L.R
\name{Transient_Process_L}
\alias{Transient_Process_L}
\title{Simulating a transient process (Loaded network)}
\usage{
Transient_Process_L(
  RealBioNet,
  ExtSelfLoop = TRUE,
  Controller = NULL,
  ConVals = NULL,
  ExternalNode = NULL,
  Times = 1000L,
  Init_1_Ratio = c(0.5, 0.5),
  NumSys = 2L,
  UpdateRule = 1L
)
}
\arguments{
\item{RealBioNet}{A four-element formatted list that records information of a Boolean network.}

\item{ExtSelfLoop}{A logical value. Should inputs be converted into
self-loops? (Default: \code{TRUE})}

\item{Controller}{An integer or character vector indicating the nodes to be
controlled. (Default: \code{NULL})}

\item{ConVals}{A Boolean vector specifying the values of controlled of
controlled nodes. (Default: \code{NULL})}

\item{ExternalNode}{A Boolean vector specifying the states of non-input nodes.
(Default: \code{NULL}, indicating random assignment)}

\item{Times}{an integer, steps of simulating system.}

\item{Init_1_Ratio}{a NumvericVector, proportion of each value (its length is 
\code{NumSys}).}

\item{NumSys}{an integer, number of discrete value.}

\item{UpdateRule}{an integer, update rule: 1=synchronous, 2=asynchronous, 
3=quick-asynchronous.}
}
\value{
A list containing two elements: 
\itemize{
\item \code{[[1]] Transient states}: A matrix whose rows and columns represent 
 time points and nodes, respectively.
\item \code{[[2]] Network}: The structure is consistent with that in 
\link{BoolGRN_CellCollective}
}
}
\description{
Same as \link{DNS_DamageSpread}, but record the transient process.
}
\details{
See \link{DNS_DamageSpread}.
}
