% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_BoolFun2Polynomial}
\alias{c_BoolFun2Polynomial}
\title{C++ prototype function for BoolFun_Polynomial()}
\usage{
c_BoolFun2Polynomial(VariableMat, MapTab, LogiSpin)
}
\arguments{
\item{VariableMat, }{a matrix that record input variables.}

\item{MapTab, }{the truth table of Boolean function.}

\item{LogiSpin, }{an integer, set the detail form of Polynomial transformation.}
}
\value{
a List, [[1]] is.SAT ture or not? [[2]] Weights of coupled variable; [[3]] threshold.
}
\description{
C++ prototype function for BoolFun_Polynomial()
}
\details{
Please see document of \link{BoolFun_Polynomial}
}
