% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_ScalingLaw_Simualtion}
\alias{c_ScalingLaw_Simualtion}
\title{C++ prototype function for DNS_Engaged()}
\usage{
c_ScalingLaw_Simualtion(
  sys_size,
  ll_system,
  obf_type,
  bias_rf,
  obf_ratio,
  net_type,
  net_f_para,
  obf_i_para1,
  obf_i_para2,
  PointedNode,
  PointValues,
  NodeDetailInfor,
  ReturnResidualNetwork
)
}
\arguments{
\item{sys_size, }{system size}

\item{ll_system, }{level number of discrete system}

\item{obf_type, }{ordered Boolean function type}

\item{bias_rf, }{biases in random function}

\item{obf_ratio, }{proportion of ordered Boolean function}

\item{net_type, }{system topological type}

\item{net_f_para, }{topological configured parameters}

\item{obf_i_para1, }{configuration parameter for OBF (1)}

\item{obf_i_para2, }{configuration parameter for OBF (2)}

\item{PointedNode, }{set which nodes should be controlled.}

\item{PointValues, }{set corresponding values of controlled genes.}

\item{NodeDetailInfor, }{should return detail informations of node.}

\item{ReturnResidualNetwork, }{should return residual networks?}
}
\description{
C++ prototype function for DNS_Engaged()
}
\details{
Please see document of \link{DNS_Engaged}
}
