% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulVFun_is_Domainted.R
\name{MulVFun_is_Domainted}
\alias{MulVFun_is_Domainted}
\title{Decode the information of a multi-valued domaineted function}
\usage{
MulVFun_is_Domainted(aVec, k, L, PrintOut = FALSE)
}
\arguments{
\item{aVec}{an IntegerVector, a mapping table of function.}

\item{k}{an integer, k-variable of function.}

\item{L}{an integer, L-level multi-valued system.}

\item{PrintOut}{a Boolean value, show output to the terminal.}
}
\value{
a 3-element list (if it is threshold): [[1]] An integer [[2]] describes 
weights of each input variable (from low to high), [[3]] records the 
L baselines of the mapping table (See examples).
}
\description{
This function can return a \code{list} records information of
whether given function is threshold, weights, baselines (deltas).
It is also compatible with Boolean system (L=2).
}
\examples{
# Show the information of a multi-valued linear threshold function.
# Return a list.
# set.seed(1002);
# A_MulV_Fun=MulVFun_Generator("D", k=2L, L=3L, MappingTable=TRUE);
# View(A_MulV_Fun); # Show the mapping table.
# A_MulV_Fun_Domaint=MulVFun_is_Domainted(A_MulV_Fun[,3], 2, 3, TRUE);
# Terminal output: 
# f(~): {w_1=-5, w_0=3 | delta_0=-1, delta_1=0, delta_2=1}
set.seed(1002)
A_MulV_Fun=MulVFun_Generator("D", k=2L, L=3L, MappingTable=TRUE)
A_MulV_Fun_Domaint=MulVFun_is_Domainted(A_MulV_Fun[,3], 2, 3, TRUE)

}
